feat: ask for whether to use taobao when get versions
