+++ packages/@vue/cli/lib/util/installDeps.js
@@ -1,77 +1,12 @@
-const request = require('./request')
 const chalk = require('chalk')
 const execa = require('execa')
 const readline = require('readline')
-const inquirer = require('inquirer')
-const { loadOptions, saveOptions } = require('../options')
-const { pauseSpinner, resumeSpinner } = require('@vue/cli-shared-utils')
+const registries = require('./registries')
+const shouldUseTaobao = require('./shouldUseTaobao')
 
 const debug = require('debug')('vue-cli:install')
 
-const registries = {
-  npm: 'https://registry.npmjs.org',
-  yarn: 'https://registry.yarnpkg.com',
-  taobao: 'https://registry.npm.taobao.org'
-}
 const taobaoDistURL = 'https://npm.taobao.org/dist'
-
-async function ping (registry) {
-  await request.get(`${registry}/vue-cli-version-marker/latest`)
-  return registry
-}
-
-function removeSlash (url) {
-  return url.replace(/\/$/, '')
-}
-
-let checked
-let result
-async function shouldUseTaobao () {
-  // ensure this only gets called once.
-  if (checked) return result
-  checked = true
-
-  // previously saved preference
-  const saved = loadOptions().useTaobaoRegistry
-  if (typeof saved === 'boolean') {
-    return (result = saved)
-  }
-
-  const save = val => {
-    result = val
-    saveOptions({ useTaobaoRegistry: val })
-    return val
-  }
-
-  const userCurrent = (await execa(`npm`, ['config', 'get', 'registry'])).stdout
-  const defaultRegistry = registries.npm
-  if (removeSlash(userCurrent) !== removeSlash(defaultRegistry)) {
-    // user has configured custom regsitry, respect that
-    return save(false)
-  }
-  const faster = await Promise.race([
-    ping(defaultRegistry),
-    ping(registries.taobao)
-  ])
-
-  if (faster !== registries.taobao) {
-    // default is already faster
-    return save(false)
-  }
-
-  // ask and save preference
-  pauseSpinner()
-  const { useTaobaoRegistry } = await inquirer.prompt([{
-    name: 'useTaobaoRegistry',
-    type: 'confirm',
-    message: chalk.yellow(
-      ` Your connection to the the default npm registry seems to be slow.\n` +
-      `   Use ${chalk.cyan(registries.taobao)} for faster installation?`
-    )
-  }])
-  resumeSpinner()
-  return save(useTaobaoRegistry)
-}
 
 function toStartOfLine (stream) {
   if (!chalk.supportsColor) {
