+++ packages/@vue/cli/lib/util/registries.js
@@ -1,0 +1,7 @@
+const registries = {
+  npm: 'https://registry.npmjs.org',
+  yarn: 'https://registry.yarnpkg.com',
+  taobao: 'https://registry.npm.taobao.org'
+}
+
+module.exports = registries
